projects/requirements.txt
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.3.2
